capstone/requirements.txt
python
django
django-annoying